refactor(actions.mail): migrate translations location to adjacent
