+++ src/routes/[[lang=lang]]/events/+page.server.ts
@@ -1,7 +1,7 @@
-import { mail } from '$server/actions/mail';
+import { mail } from '$shared/actions/mail/mail.server';
+import { translations as mailT } from '$shared/actions/mail/translation';
 import { createMockedEvents } from '$shared/mocks';
 import type { Language } from '$shared/services/i18n';
-import { translations as mailT } from '$shared/services/i18n/translations/actions/mail';
 import { translations as pageT } from '$shared/services/i18n/translations/pages/events';
 
 import type { PageServerLoadEvent, Actions } from './$types';
